+++ src/addin/client-interfaces/addin-client-open-help-args.ts
@@ -5,7 +5,6 @@ export interface AddinClientOpenHelpArgs {
 
   /**
    * The name of the help key to be launched.
-   * @deprecated See _Deprecated help support_ in README
    */
   helpKey: string;