+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.5.0 (2024-10-07)
+- Added `showInlineHelp`, `helpPopoverContent`, and `helpPopoverTitle` properties in `AddinTileConfig` to support inline help in tile header.
+- Added `inlineHelpClick` callback in `AddinClientCallbacks` to handle inline help button click.
+
 # 1.4.1 (2024-09-10)
 - Addressed issue to prevent `supportedEventTypes` to be `undefined`. Should default to empty array.